tests/wpt/web-platform-tests/shadow-dom/untriaged/elements-and-dom-objects/shadowroot-object/shadowroot-attributes/test-014.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: ShadowRoot olderShadowRoot</title>
<link rel="author" title="Kenji Baheux" href="mailto:kenjibaheux@google.com">
<link rel="author" title="Kyohei Tsukuda" href="tsukuda.kyouhei@gmail.com">
<link rel="help" href="http://w3c.github.io/webcomponents/spec/shadow/#widl-ShadowRoot-olderShadowRoot">
<meta name="assert" content="The ShadowRoot element: olderShadowRoot attribute">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
test(unit(function (ctx) {

	var d = newRenderedHTMLDocument(ctx);

	var host = d.createElement('div');
	d.body.appendChild(host);

	// Shadow root and older Shadow root to play with
	var oldRoot = host.createShadowRoot();

	var div = d.createElement('div');
	div.innerHTML = '' +
		'<span id="spandex">This is a shadow root content</span>' +
		'<content><span id="contentId">This is the content fallback</span></content>';
	oldRoot.appendChild(div);
	var youngRoot = host.createShadowRoot();

	assert_equals(oldRoot.olderShadowRoot, null, 'If the context object is the oldest shadow root, return null');
	assert_equals(youngRoot.olderShadowRoot, oldRoot, 'Return the older shadow root relative to the context object');

}), 'ShadowRoot.olderShadowRoot_T01');
</script>
</body>
</html>
